// frontend/src/app/core/api/generated/model/product-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductResponse { 
    id?: string;
    sku?: string;
    name?: string;
    description?: string;
    price?: number;
    active?: boolean;
    createdAt?: string;
    updatedAt?: string;
}